Методы для типа переплета перенесены в компонент CoverTypes в папке API
